backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.27
google-generativeai==0.5.4
Pillow==10.3.0
pydantic==2.7.4
python-dotenv==1.0.1
cors==1.0.1
